# Cache `/api/-/search` results