--- a/server/src/main/java/org/eclipse/openvsx/cache/CacheService.java
+++ b/server/src/main/java/org/eclipse/openvsx/cache/CacheService.java
@@ -12,11 +12,14 @@ package org.eclipse.openvsx.cache;
 import org.eclipse.openvsx.entities.Extension;
 import org.eclipse.openvsx.entities.ExtensionVersion;
 import org.eclipse.openvsx.entities.UserData;
+import org.eclipse.openvsx.json.SearchEntryJson;
 import org.eclipse.openvsx.repositories.RepositoryService;
+import org.eclipse.openvsx.search.ExtensionSearch;
 import org.eclipse.openvsx.util.TargetPlatform;
 import org.eclipse.openvsx.util.VersionAlias;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.CacheManager;
+import org.springframework.data.elasticsearch.core.SearchHit;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
@@ -27,6 +30,7 @@ public class CacheService {
 
     public static final String CACHE_DATABASE_SEARCH = "database.search";
     public static final String CACHE_EXTENSION_JSON = "extension.json";
+    public static final String CACHE_SEARCH_ENTRY_JSON = "search.entry.json";
     public static final String CACHE_LATEST_EXTENSION_VERSION = "latest.extension.version";
     public static final String CACHE_NAMESPACE_DETAILS_JSON = "namespace.details.json";
     public static final String CACHE_AVERAGE_REVIEW_RATING = "average.review.rating";
@@ -42,6 +46,9 @@ public class CacheService {
 
     @Autowired
     ExtensionJsonCacheKeyGenerator extensionJsonCacheKey;
+
+    @Autowired
+    SearchEntryJsonCacheKeyGenerator searchEntryJsonCacheKeyGenerator;
 
     @Autowired
     LatestExtensionVersionCacheKeyGenerator latestExtensionVersionCacheKey;
@@ -89,6 +96,33 @@ public class CacheService {
             for(var targetPlatform : targetPlatforms) {
                 cache.evictIfPresent(extensionJsonCacheKey.generate(namespaceName, extensionName, targetPlatform, version));
             }
+        }
+    }
+
+    public SearchEntryJson getSearchEntryJson(SearchHit<ExtensionSearch> searchHit, boolean includeAllVersions) {
+        var cache = cacheManager.getCache(CACHE_SEARCH_ENTRY_JSON);
+        return cache != null
+                ? cache.get(searchEntryJsonCacheKeyGenerator.generate(searchHit.getContent().id, includeAllVersions), SearchEntryJson.class)
+                : null;
+    }
+
+    public void putSearchEntryJson(SearchEntryJson searchEntry, SearchHit<ExtensionSearch> searchHit, boolean includeAllVersions) {
+        var cache = cacheManager.getCache(CACHE_SEARCH_ENTRY_JSON);
+        if(cache != null) {
+            cache.put(searchEntryJsonCacheKeyGenerator.generate(searchHit.getContent().id, includeAllVersions), searchEntry);
+        }
+    }
+
+    public void evictSearchEntryJsons(Extension extension) {
+        var cache = cacheManager.getCache(CACHE_SEARCH_ENTRY_JSON);
+        if(cache == null) {
+            return; // cache is not created
+        }
+
+        var includeAllVersionsList = List.of(true, false);
+        for(var includeAllVersions : includeAllVersionsList) {
+            var key = searchEntryJsonCacheKeyGenerator.generate(extension.getId(), includeAllVersions);
+            cache.evictIfPresent(key);
         }
     }