--- a/server/src/main/java/org/eclipse/openvsx/entities/Extension.java
+++ b/server/src/main/java/org/eclipse/openvsx/entities/Extension.java
@@ -62,6 +62,7 @@ public class Extension implements Serializable {
         search.name = this.getName();
         search.namespace = this.getNamespace().getName();
         search.extensionId = search.namespace + "." + search.name;
+        search.averageRating = this.getAverageRating();
         search.downloadCount = this.getDownloadCount();
         search.targetPlatforms = this.getVersions().stream()
                 .map(ExtensionVersion::getTargetPlatform)