--- a/server/src/main/resources/ehcache.xml
+++ b/server/src/main/resources/ehcache.xml
@@ -30,6 +30,16 @@
             <heap unit="entries">1024</heap>
         </resources>
     </cache>
+    <cache alias="search.entry.json">
+        <expiry>
+            <ttl unit="seconds">3600</ttl>
+        </expiry>
+        <resources>
+            <heap unit="entries">1024</heap>
+            <offheap unit="MB">32</offheap>
+            <disk unit="MB">128</disk>
+        </resources>
+    </cache>
     <cache alias="extension.json">
         <expiry>
             <ttl unit="seconds">3600</ttl>