### Performance Comparison

<pre>
== MASTER <a href="https://github.com/amvanbaren/openvsx/commit/f9561102728657fc301c274ec0c4b00b57f09f08">f956110</a> 3 RUNS ===============================================================
---- Global Information --------------------------------------------------------
> request count                                       5000 (OK=5000   KO=0     )
> min response time                                      4 (OK=4      KO=-     )
> max response time                                   2193 (OK=2193   KO=-     )
> mean response time                                    57 (OK=57     KO=-     )
> std deviation                                        106 (OK=106    KO=-     )
> response time 50th percentile                         20 (OK=20     KO=-     )
> response time 75th percentile                         39 (OK=39     KO=-     )
> response time 95th percentile                        342 (OK=342    KO=-     )
> response time 99th percentile                        446 (OK=446    KO=-     )
> mean requests/sec                                 <b>83.333</b> (OK=83.333 KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          4997 (100%)
> 800 ms < t < 1200 ms                                   1 (  0%)
> t > 1200 ms                                            2 (  0%)
> failed                                                 0 (  0%)
================================================================================
</pre>

<pre>
== NEW 1 RUN ===================================================================
---- Global Information --------------------------------------------------------
> request count                                       5000 (OK=5000   KO=0     )
> min response time                                      4 (OK=4      KO=-     )
> max response time                                   6210 (OK=6210   KO=-     )
> mean response time                                    44 (OK=44     KO=-     )
> std deviation                                        178 (OK=178    KO=-     )
> response time 50th percentile                         20 (OK=20     KO=-     )
> response time 75th percentile                         39 (OK=39     KO=-     )
> response time 95th percentile                        114 (OK=114    KO=-     )
> response time 99th percentile                        350 (OK=350    KO=-     )
> mean requests/sec                                <b>104.167</b> (OK=104.167 KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          4983 (100%)
> 800 ms < t < 1200 ms                                   3 (  0%)
> t > 1200 ms                                           14 (  0%)
> failed                                                 0 (  0%)
================================================================================
</pre>

<pre>
== NEW 3 RUNS ==================================================================
---- Global Information --------------------------------------------------------
> request count                                       5000 (OK=5000   KO=0     )
> min response time                                      3 (OK=3      KO=-     )
> max response time                                    407 (OK=407    KO=-     )
> mean response time                                    18 (OK=18     KO=-     )
> std deviation                                         25 (OK=25     KO=-     )
> response time 50th percentile                         11 (OK=11     KO=-     )
> response time 75th percentile                         17 (OK=17     KO=-     )
> response time 95th percentile                         53 (OK=53     KO=-     )
> response time 99th percentile                        139 (OK=139    KO=-     )
> mean requests/sec                                    <b>250</b> (OK=250    KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          5000 (100%)
> 800 ms < t < 1200 ms                                   0 (  0%)
> t > 1200 ms                                            0 (  0%)
> failed                                                 0 (  0%)
================================================================================
</pre>