@filiptronicek 2022-10-03
@amvanbaren could you please share what you used to get this benchmark so I could try to reproduce? 🙏