@filiptronicek 2022-10-03
Made a script for downloading a lot of extensions: https://gist.github.com/filiptronicek/fdfc009371401b9d63716759a90887f4 (should be easily tweak-able for changing download amount as well as output directory).