@amvanbaren 2023-01-18
Issue #379 comes close. I've added a priority.