@amvanbaren 2023-03-31
I've tested this on staging. It looks like the load balancer is doing the heavy lifting, because the results are more or less the same. It can still be useful for #670 though.

<pre>
== CURRENT =====================================================================
---- Global Information --------------------------------------------------------
> request count                                       5000 (OK=4244   KO=756   )
> min response time                                    141 (OK=141    KO=144   )
> max response time                                   1507 (OK=1507   KO=502   )
> mean response time                                   175 (OK=178    KO=161   )
> std deviation                                         97 (OK=104    KO=37    )
> response time 50th percentile                        150 (OK=150    KO=155   )
> response time 75th percentile                        157 (OK=156    KO=162   )
> response time 95th percentile                        305 (OK=312    KO=176   )
> response time 99th percentile                        627 (OK=734    KO=459   )
> mean requests/sec                                 <b>27.933</b> (OK=23.709 KO=4.223 )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          4221 ( 84%)
> 800 ms < t < 1200 ms                                  22 (  0%)
> t > 1200 ms                                            1 (  0%)
> failed                                               756 ( 15%)
---- Errors --------------------------------------------------------------------
> status.find.in(200,201,202,203,204,205,206,207,208,209,304), f    581 (76,85%)
ound 400
> status.find.in(200,201,202,203,204,205,206,207,208,209,304), f    175 (23,15%)
ound 429
================================================================================
</pre>

<pre>
== SEARCH CACHE ================================================================
---- Global Information --------------------------------------------------------
> request count                                       5000 (OK=4244   KO=756   )
> min response time                                    142 (OK=142    KO=144   )
> max response time                                   1962 (OK=1962   KO=611   )
> mean response time                                   184 (OK=188    KO=158   )
> std deviation                                        128 (OK=138    KO=35    )
> response time 50th percentile                        150 (OK=149    KO=152   )
> response time 75th percentile                        156 (OK=155    KO=159   )
> response time 95th percentile                        448 (OK=456    KO=172   )
> response time 99th percentile                        761 (OK=870    KO=308   )
> mean requests/sec                                 <b>26.738</b> (OK=22.695 KO=4.043 )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          4197 ( 84%)
> 800 ms < t < 1200 ms                                  37 (  1%)
> t > 1200 ms                                           10 (  0%)
> failed                                               756 ( 15%)
---- Errors --------------------------------------------------------------------
> status.find.in(200,201,202,203,204,205,206,207,208,209,304), f    568 (75,13%)
ound 400
> status.find.in(200,201,202,203,204,205,206,207,208,209,304), f    188 (24,87%)
ound 429
================================================================================
</pre>